+++ src/Headroom/FileSupport.hs
@@ -36,6 +36,7 @@ import qualified Headroom.FileSupport.Haskell       as Haskell
 import qualified Headroom.FileSupport.HTML          as HTML
 import qualified Headroom.FileSupport.Java          as Java
 import qualified Headroom.FileSupport.JS            as JS
+import qualified Headroom.FileSupport.Kotlin        as Kotlin
 import qualified Headroom.FileSupport.PHP           as PHP
 import qualified Headroom.FileSupport.PureScript    as PureScript
 import qualified Headroom.FileSupport.Rust          as Rust
@@ -66,6 +67,7 @@ fileSupport Haskell    = Haskell.fileSupport
 fileSupport HTML       = HTML.fileSupport
 fileSupport Java       = Java.fileSupport
 fileSupport JS         = JS.fileSupport
+fileSupport Kotlin     = Kotlin.fileSupport
 fileSupport PHP        = PHP.fileSupport
 fileSupport PureScript = PureScript.fileSupport
 fileSupport Rust       = Rust.fileSupport