+++ src/Headroom/FileType.hs
@@ -66,6 +66,7 @@ configByFileType HeadersConfig {..} fileType = case fileType of
   HTML       -> hscHtml
   Java       -> hscJava
   JS         -> hscJs
+  Kotlin     -> hscKotlin
   PHP        -> hscPhp
   PureScript -> hscPureScript
   Rust       -> hscRust