+++ src/Headroom/FileType/Types.hs
@@ -30,6 +30,7 @@ data FileType
   | HTML       -- ^ support for /HTML/
   | Java       -- ^ support for /Java/ programming language
   | JS         -- ^ support for /JavaScript/ programming language
+  | Kotlin     -- ^ support for /Kotlin/ programming language
   | PHP        -- ^ support for /PHP/ programming language
   | PureScript -- ^ support for /PureScript/ programming language
   | Rust       -- ^ support for /Rust/ programming language