+++ test-data/code-samples/kotlin/sample1.kt
@@ -1,0 +1,10 @@
+/*
+ * This is copyright header
+ */
+
+package foo
+
+/** this is not license header */
+fun main(args : Array<String>) {
+    println("Hello, World!")
+}